+++ tests/layers_views/layers/ImageLayer.spec.js
@@ -60,7 +60,7 @@ test('ImageLayer', t => {
     target: [2, 2, 0],
     zoom: 0
   });
-  let state = { data: [], width: 0, height: 0 };
+  const state = { data: [], width: 0, height: 0 };
   const testCases = [
     {
       props: {
@@ -75,7 +75,7 @@ test('ImageLayer', t => {
         channelIsOn: [true, false],
         loader: {
           getRaster: async () => {
-            state.data.push(new Uint32Array([0, 2, 1, 2]))
+            state.data.push(new Uint32Array([0, 2, 1, 2]));
             state.width = 2;
             state.width = 2;
           },